Implement point style for line geometries
